+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineSession.kt
@@ -24,9 +24,6 @@ import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.Settings
 import mozilla.components.concept.engine.history.HistoryTrackingDelegate
 import mozilla.components.concept.engine.request.RequestInterceptor
-import mozilla.components.concept.engine.shopping.ProductAnalysis
-import mozilla.components.concept.engine.shopping.ProductAnalysisStatus
-import mozilla.components.concept.engine.shopping.ProductRecommendation
 import mozilla.components.concept.engine.translate.TranslationOptions
 import org.json.JSONObject
 import kotlin.reflect.KProperty
@@ -434,96 +431,6 @@ class SystemEngineSession(
         onException: (Throwable) -> Unit,
     ) {
         throw UnsupportedOperationException("Getting web compat info is not available in this engine")
-    }
-
-    /**
-     * /**
-     * See [EngineSession.requestProductRecommendations]
-     */
-     */
-    override fun requestProductRecommendations(
-        url: String,
-        onResult: (List<ProductRecommendation>) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Analysis of product reviews for shopping is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.requestProductAnalysis]
-     */
-    override fun requestProductAnalysis(
-        url: String,
-        onResult: (ProductAnalysis) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Analysis of product reviews for shopping is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.reanalyzeProduct]
-     */
-    override fun reanalyzeProduct(
-        url: String,
-        onResult: (String) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Reanalyzing product reviews for shopping is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.requestAnalysisStatus]
-     */
-    override fun requestAnalysisStatus(
-        url: String,
-        onResult: (ProductAnalysisStatus) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Requesting product analysis status is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.sendClickAttributionEvent]
-     */
-    override fun sendClickAttributionEvent(
-        aid: String,
-        onResult: (Boolean) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Sending click attribution event is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.sendImpressionAttributionEvent]
-     */
-    override fun sendImpressionAttributionEvent(
-        aid: String,
-        onResult: (Boolean) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Sending impression attribution event is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.sendPlacementAttributionEvent]
-     */
-    override fun sendPlacementAttributionEvent(
-        aid: String,
-        onResult: (Boolean) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Sending placement attribution event is not available in this engine")
-    }
-
-    /**
-     * See [EngineSession.reportBackInStock]
-     */
-    override fun reportBackInStock(
-        url: String,
-        onResult: (String) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {
-        throw UnsupportedOperationException("Reporting back in stock is not available in this engine")
     }
 
     /**