+++ mobile/android/android-components/components/browser/state/src/test/java/mozilla/components/browser/state/engine/EngineObserverTest.kt
@@ -38,9 +38,6 @@ import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.mediasession.MediaSession
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
-import mozilla.components.concept.engine.shopping.ProductAnalysis
-import mozilla.components.concept.engine.shopping.ProductAnalysisStatus
-import mozilla.components.concept.engine.shopping.ProductRecommendation
 import mozilla.components.concept.engine.translate.TranslationError
 import mozilla.components.concept.engine.translate.TranslationOperation
 import mozilla.components.concept.engine.translate.TranslationOptions
@@ -88,46 +85,6 @@ class EngineObserverTest {
             ) {}
             override fun getWebCompatInfo(
                 onResult: (JSONObject) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestProductAnalysis(
-                url: String,
-                onResult: (ProductAnalysis) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestProductRecommendations(
-                url: String,
-                onResult: (List<ProductRecommendation>) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun reanalyzeProduct(
-                url: String,
-                onResult: (String) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestAnalysisStatus(
-                url: String,
-                onResult: (ProductAnalysisStatus) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendClickAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendImpressionAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendPlacementAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun reportBackInStock(
-                url: String,
-                onResult: (String) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
             override fun requestTranslate(
@@ -214,46 +171,6 @@ class EngineObserverTest {
                 onResult: (JSONObject) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
-            override fun requestProductAnalysis(
-                url: String,
-                onResult: (ProductAnalysis) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestProductRecommendations(
-                url: String,
-                onResult: (List<ProductRecommendation>) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun reanalyzeProduct(
-                url: String,
-                onResult: (String) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestAnalysisStatus(
-                url: String,
-                onResult: (ProductAnalysisStatus) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendClickAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendImpressionAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendPlacementAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun reportBackInStock(
-                url: String,
-                onResult: (String) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
             override fun requestTranslate(
                 fromLanguage: String,
                 toLanguage: String,
@@ -332,46 +249,6 @@ class EngineObserverTest {
             ) {}
             override fun getWebCompatInfo(
                 onResult: (JSONObject) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestProductRecommendations(
-                url: String,
-                onResult: (List<ProductRecommendation>) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestProductAnalysis(
-                url: String,
-                onResult: (ProductAnalysis) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun reanalyzeProduct(
-                url: String,
-                onResult: (String) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun requestAnalysisStatus(
-                url: String,
-                onResult: (ProductAnalysisStatus) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendClickAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendImpressionAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun sendPlacementAttributionEvent(
-                aid: String,
-                onResult: (Boolean) -> Unit,
-                onException: (Throwable) -> Unit,
-            ) {}
-            override fun reportBackInStock(
-                url: String,
-                onResult: (String) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
             override fun requestTranslate(